Distributed Pi/requirements.txt
Flask==3.0.0
influxdb-client==1.38.0
paho-mqtt==1.6.1
flask-cors==4.0.0
flask-socketio==5.3.6